requirements.txt
asgiref==3.6.0
asttokens==2.2.1
autopep8==2.0.1
av==10.0.0
backcall==0.2.0
blis==0.7.9
cachetools==5.3.0
catalogue==2.0.8
certifi==2022.12.7
chardet==3.0.4
charset-normalizer==3.0.1
classy-classification==0.6.2
click==8.1.3
comm==0.1.2
confection==0.0.4
cymem==2.0.7
debugpy==1.6.6
decorator==5.1.1
Django==4.1.7
djangorestframework==3.14.0
easyocr==1.6.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl
executing==1.2.0
filelock==3.9.0
google-api-core==2.11.0
google-api-python-client==2.78.0
google-auth==2.16.1
google-auth-httplib2==0.1.0
google-trans-new==1.1.9
googleapis-common-protos==1.58.0
googletrans==3.1.0a0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httplib2==0.21.0
httpx==0.13.3
huggingface-hub==0.12.1
hyperframe==5.2.0
idna==2.10
imageio==2.25.1
ipykernel==6.21.2
ipython==8.10.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jupyter_client==8.0.3
jupyter_core==5.2.0
langcodes==3.3.0
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
murmurhash==1.0.9
nest-asyncio==1.5.6
networkx==3.0
ninja==1.11.1
nltk==3.8.1
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
opencv-python==4.5.5.64
opencv-python-headless==4.5.4.60
packaging==23.0
pandas==1.5.3
parso==0.8.3
pathy==0.10.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.0.0
preshed==3.0.8
prompt-toolkit==3.0.36
protobuf==4.22.0
psutil==5.9.4
psycopg-binary==3.1.8
psycopg2==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyclipper==1.3.0.post4
pycodestyle==2.10.0
pydantic==1.10.5
Pygments==2.14.0
pyparsing==3.0.9
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2022.7.1
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.0.0
regex==2022.10.31
requests==2.28.2
rfc3986==1.5.0
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.2.1
scipy==1.10.0
sentence-transformers==2.2.2
sentencepiece==0.1.97
shapely==2.0.1
six==1.16.0
smart-open==6.3.0
sniffio==1.3.0
spacy==3.5.0
spacy-alignments==0.9.0
spacy-legacy==3.0.12
spacy-loggers==1.0.4
spacy-transformers==1.2.2
sqlparse==0.4.3
srsly==2.4.5
stack-data==0.6.2
thinc==8.1.7
threadpoolctl==3.1.0
tifffile==2023.2.3
tokenizers==0.13.2
tomli==2.0.1
torch==1.13.1
torchvision==0.14.1
tornado==6.2
tqdm==4.64.1
traitlets==5.9.0
transformers==4.26.1
typer==0.7.0
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.14
wasabi==1.1.1
wcwidth==0.2.6
